[buf fixed] change correct type to float

+++ train.py
@@ -145,7 +145,7 @@ def eval_training(epoch):
 
     #add informations to tensorboard
     writer.add_scalar('Test/Average loss', test_loss / len(cifar100_test), epoch)
-    writer.add_scalar('Test/Accuracy', correct / len(cifar100_test), epoch)
+    writer.add_scalar('Test/Accuracy', correct.float() / len(cifar100_test), epoch)
 
     return correct / len(cifar100_test)
